+++ Pod/Classes/Public/MP.m
@@ -21,7 +21,7 @@
 #import <CoreFoundation/CoreFoundation.h>
 #import "MPLayoutFactory.h"
 
-NSString * const kMPLibraryVersion = @"3.0.4";
+NSString * const kMPLibraryVersion = @"3.0.5";
 
 NSString * const kLaterActionIdentifier = @"LATER_ACTION_IDENTIFIER";
 NSString * const kPrintActionIdentifier = @"PRINT_ACTION_IDENTIFIER";